% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_closeAllConnections.R
\name{trace_closeAllConnections}
\alias{trace_closeAllConnections}
\title{Trace all calls to closeAllConnections()}
\usage{
trace_closeAllConnections(
  action = c("error", "warning"),
  allow = list(base::sys.save.image),
  enable = TRUE,
  calls = TRUE
)
}
\arguments{
\item{action}{(character) What action to take when called.}

\item{allow}{(list) List of functions that are allowed to call
\code{closeAllConnections()} and that this tracer will ignore.}

\item{enable}{(logical) Enable or disable tracing.}

\item{calls}{(logical) If TRUE, the call stack is part of the
condition message, otherwise not.}
}
\value{
Nothing.
}
\description{
Trace all calls to closeAllConnections()
}
\details{
Calling \code{\link[base:showConnections]{base::closeAllConnections()}} is a very harsh action.  It closes
all open R connections regardless who created them.  It should never be
called in an R package.  There's also little reason for calling it from
an R script, or even from the R prompt.  If you find yourself having to
do so, your R session is probably already in such a bad shape that it's
better to restart R all along.

This tracer detects whenever \code{closeAllConnections()} is called.
It can then produce an informative warning or and error.  If an error,
then no connections will be closed.
}
