% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_detectCores.R
\name{trace_detectCores}
\alias{trace_detectCores}
\title{Trace all calls to parallel::detectCores()}
\usage{
trace_detectCores(
  action = c("warning", "error"),
  allow = list(base::sys.save.image),
  enable = TRUE,
  calls = TRUE
)
}
\arguments{
\item{action}{(character) What action to take when called.}

\item{allow}{(list) List of functions that are allowed to call
\code{parallel::detectCores()} and that this tracer will ignore.}

\item{enable}{(logical) Enable or disable tracing.}

\item{calls}{(logical) If TRUE, the call stack is part of the
condition message, otherwise not.}
}
\value{
Nothing.

This tracer detects whenever \code{parallel::detectCores()} is called.
It can then produce an informative warning or and error.  If an error,
then no connections will be closed.
}
\description{
Trace all calls to parallel::detectCores()
}
