% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_rm.R
\name{trace_rm}
\alias{trace_rm}
\title{Trace all rm(list = ls()) calls}
\usage{
trace_rm(action = c("error", "warning"), enable = TRUE)
}
\arguments{
\item{action}{(character) What action to take when called.}

\item{enable}{(logical) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Trace all rm(list = ls()) calls
}
\details{
Calling \code{rm(list = ls())} is considered bad manners.

This tracer detects whenever \code{rm(list = ls())} is called.
It can then produce an informative warning or and error.  If an error,
then the call is not completed.
}
