% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_rng_on_load.R
\name{trace_rng_on_load}
\alias{trace_rng_on_load}
\title{Trace packages that update the random number generation (RNG) state when loaded}
\usage{
trace_rng_on_load(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Trace packages that update the random number generation (RNG) state when loaded
}
\details{
This tracer injects itself as a tracer to \code{\link[base:ns-load]{base::loadNamespace()}}, where
it keeps track of the \code{.Random.seed} in the global environment as packages
are \emph{loaded}.  This way it can detect if a package updates the random
number generation (RNG) state during load, e.g. when the package's
\code{.onLoad()} function is called.
Importantly, currently, it does not detect RNG-state changes when a package
is \emph{attached}, e.g. when the package's \code{.onAttach()} function is called.
}
