% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_connections.R
\name{track_connections}
\alias{track_connections}
\title{Warn when the connections are opened or closed}
\usage{
track_connections(action = c("on", "off", "init"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when the connections are opened or closed
}
\examples{
\dontrun{
track_connections()
}

}
