% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_envvars.R
\name{track_envvars}
\alias{track_envvars}
\title{Warn when environment variables have been changed}
\usage{
track_envvars(action = c("on", "off", "init"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when environment variables have been changed
}
\examples{
\dontrun{
track_envvars()
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
