% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_gc.R
\name{track_gc}
\alias{track_gc}
\title{Warn when garbage collector has been run}
\usage{
track_gc(action = c("on", "off", "init"))
}
\arguments{
\item{action}{(character) Enable or disable tracking.}
}
\value{
Nothing.
}
\description{
Warn when garbage collector has been run
}
\examples{
\dontrun{
track_gc()
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
