% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_globalenv.R
\name{track_globalenv}
\alias{track_globalenv}
\title{Warn when objects are added or removed from the global environment}
\usage{
track_globalenv(action = c("on", "off", "init"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when objects are added or removed from the global environment
}
\details{
Set R option \option{tracker.globalenv} to \code{FALSE} to disable.

Global variables that are ignored:
\itemize{
\item \code{.Random.seed}
}
}
\examples{
\dontrun{
track_globalenv()
}

}
