% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_options.R
\name{track_options}
\alias{track_options}
\title{Warn when certain R options have been changed}
\usage{
track_options(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when certain R options have been changed
}
\details{
Options that are monitored:
\itemize{
\item \code{stringsAsFactors}
\item \code{contrasts}
\item \code{na.action}
\item \code{ts.eps}
\item \code{ts.S.compat}
}
}
\examples{
\dontrun{
track_options()
}

}
