% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_packages.R
\name{track_packages}
\alias{track_packages}
\title{Warn when packages are loaded or unloaded}
\usage{
track_packages(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when packages are loaded or unloaded
}
\details{
:
Set R option \option{tracker.packages} to \code{FALSE} to disable.
}
\examples{
\dontrun{
track_packages()
}

}
