% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_rng.R
\name{track_rng}
\alias{track_rng}
\title{Warn when the .Random.seed has changed}
\usage{
track_rng()
}
\description{
Warn when the .Random.seed has changed
}
\details{
Global variables that are monitored:
\itemize{
\item \code{.Random.seed}
}
}
\examples{
\dontrun{
tracker_rng()
}

}
