% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_rplots_files.R
\name{track_rplots_files}
\alias{track_rplots_files}
\title{Alert about newly produced Rplots*.pdf files}
\usage{
track_rplots_files(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Look for newly produced \code{Rplots*.pdf} files, which may be produced
when running in batch mode or when screen devices are not available.
See \code{?options} and option \code{'device'}.
}
\examples{
\dontrun{
track_rplots_files()
}

}
