% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_sinks.R
\name{track_sinks}
\alias{track_sinks}
\title{Warn when the output or message stream is sinked}
\usage{
track_sinks(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when the output or message stream is sinked
}
\details{
The behavior of this tracker can be configured via R option
\option{tracker.sink.delta};
\itemize{
\item \code{TRUE}: Report only when there is a change in the sinks
\item \code{FALSE}: (default) Report whenever there is an active sink
}
}
\examples{
\dontrun{
track_sinks()
}

}
