% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_time.R
\name{track_time}
\alias{track_time}
\title{Report on processing time}
\usage{
track_time(threshold = 0, action = c("on", "off", "init"))
}
\arguments{
\item{threshold}{(numeric) Minimum processing time (in seconds) to
report on.}

\item{action}{(character) Enable or disable tracking.}
}
\value{
Nothing.
}
\description{
Report on processing time
}
\examples{
\dontrun{
track_time(threshold = 0.2) # track anything greater than 0.2s
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
