% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_files.R
\name{tracker_files}
\alias{tracker_files}
\title{Warn when files are added or removed}
\usage{
tracker_files(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Warn when files are added or removed
}
\details{
Set R option \option{tracker.files} to \code{FALSE} to disable.
}
\examples{
\dontrun{
addTaskCallback(tracker_globals, name = "Files tracker")
}

}
