% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_files.R, R/tracker_globals.R
\name{tracker_globals}
\alias{tracker_globals}
\title{Warn when files are added or removed}
\usage{
tracker_globals(expr, value, ok, visible)

tracker_globals(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE

Always TRUE
}
\description{
Warn when files are added or removed

Warn when objects are added or removed from the global environment
}
\details{
Set R option \option{tracker.files} to \code{FALSE} to disable.

Set R option \option{tracker.globals} to \code{FALSE} to disable.

Global variables that are ignored:
\itemize{
\item \code{.Random.seed}
}
}
\examples{
\dontrun{
addTaskCallback(tracker_globals, name = "Files tracker")
}

\dontrun{
addTaskCallback(tracker_globals, name = ".GlobalEnv tracker")
}

}
