% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_locale.R
\name{tracker_locale}
\alias{tracker_locale}
\title{Warn when locale have been changed}
\usage{
tracker_locale(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Warn when locale have been changed
}
\examples{
\dontrun{
addTaskCallback(tracker_locale, name = "Locale tracker")
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
