% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_options.R
\name{tracker_options}
\alias{tracker_options}
\title{Warn when certain R options have been changed}
\usage{
tracker_options(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Warn when certain R options have been changed
}
\details{
Options that are monitored:
\itemize{
\item \code{stringsAsFactors}
\item \code{contrasts}
\item \code{na.action}
\item \code{ts.eps}
\item \code{ts.S.compat}
}
}
\examples{
\dontrun{
addTaskCallback(tracker_options, name = "Options tracker")
}

}
