% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_package.R
\name{tracker_package}
\alias{tracker_package}
\title{Warn when packages are loaded or unloaded}
\usage{
tracker_package(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE
}
\description{
Warn when packages are loaded or unloaded
}
\details{
:
Set R option \option{tracker.packages} to \code{FALSE} to disable.
}
\examples{
\dontrun{
addTaskCallback(tracker_package, name = "Package tracker")
}

}
