% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_rng.R
\name{tracker_rng}
\alias{tracker_rng}
\title{Warn when the .Random.seed has changed}
\usage{
tracker_rng(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (As of R 4.2.1, this is
\link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE
}
\description{
Warn when the .Random.seed has changed
}
\details{
Global variables that are monitored:
\itemize{
\item \code{.Random.seed}
}
}
