% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_rplot.R
\name{tracker_rplots}
\alias{tracker_rplots}
\title{Alert about newly produced Rplots*.pdf files}
\usage{
tracker_rplots(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (As of R 4.2.1, this is
\link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE
}
\description{
Look for newly produced \code{Rplots*.pdf} files, which may be produced
when running in batch mode or when screen devices are not available.
See \code{?options} and option \code{'device'}.
}
