% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_rplots_files.R
\name{tracker_rplots_files}
\alias{tracker_rplots_files}
\title{Alert about newly produced Rplots*.pdf files}
\usage{
tracker_rplots_files(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Look for newly produced \code{Rplots*.pdf} files, which may be produced
when running in batch mode or when screen devices are not available.
See \code{?options} and option \code{'device'}.
}
\examples{
\dontrun{
addTaskCallback(tracker_rplots_files, name = "Rplots tracker")
}

}
