% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_sink.R
\name{tracker_sink}
\alias{tracker_sink}
\title{Warn when the output or message stream is sinked}
\usage{
tracker_sink(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE
}
\description{
Warn when the output or message stream is sinked
}
\details{
The behavior of this tracker can be configured via R option
\option{tracker.sink.delta};
\itemize{
\item \code{TRUE}: Report only when there is a change in the sinks
\item \code{FALSE}: (default) Report whenever there is an active sink
}
}
\examples{
\dontrun{
addTaskCallback(tracker_sink, name = "Sink tracker")
}

}
