% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_files.R
\name{track_files}
\alias{track_files}
\title{Warn when files are added or removed}
\usage{
track_files(action = c("on", "off"))
}
\arguments{
\item{action}{(character) Enable or disable tracing.}
}
\value{
Nothing.
}
\description{
Warn when files are added or removed
}
\details{
Set R option \option{tracker.files} to \code{FALSE} to disable.
}
\examples{
\dontrun{
track_files()
}

}
