% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_gc.R
\name{track_gc}
\alias{track_gc}
\title{Report on garbage collection}
\usage{
track_gc(threshold = 0, action = c("on", "off", "init"))
}
\arguments{
\item{threshold}{(numeric) Minimum garbage-collection time (in seconds) to
report on.}

\item{action}{(character) Enable or disable tracking.}
}
\value{
Nothing.
}
\description{
Report on garbage collection
}
\examples{
\dontrun{
track_gc()
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
