% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_locale.R
\name{track_locale}
\alias{track_locale}
\title{Warn when locale have been changed}
\usage{
track_locale()
}
\description{
Warn when locale have been changed
}
\examples{
\dontrun{
tracker_locale()
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
