% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_time.R
\name{track_time}
\alias{track_time}
\title{Warn when processing time ellapsed}
\usage{
track_time(action = c("on", "off", "init"))
}
\arguments{
\item{action}{(character) Enable or disable tracking.}
}
\value{
Nothing.
}
\description{
Warn when processing time ellapsed
}
\examples{
\dontrun{
track_time()
}

}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
