% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_envvars.R
\name{tracker_envvars}
\alias{tracker_envvars}
\title{Warn when environment variables have been changed}
\usage{
tracker_envvars(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (As of R 4.2.1, this is
\link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Warn when environment variables have been changed
}
\seealso{
\code{\link[base:taskCallback]{base::addTaskCallback()}}
}
