% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_globals.R
\name{tracker_globals}
\alias{tracker_globals}
\title{Warn when objects are added or removed from the global environment}
\usage{
tracker_globals(expr, value, ok, visible)
}
\arguments{
\item{expr}{The \R \link[base:expression]{expression} called.}

\item{value}{The result of the evaluated expression.}

\item{ok}{A logical indicating whether it was successfully completed
or not. (This is \link[base:addTaskCallback]{always TRUE at present}.)}

\item{visible}{A logical indicating whether the result was printed or not.}
}
\value{
Always TRUE, which is used to tell \code{\link[base:taskCallback]{base::addTaskCallback()}} to keep
this callback tracker function after it's called.
}
\description{
Warn when objects are added or removed from the global environment
}
\details{
Set R option \option{tracker.globals} to \code{FALSE} to disable.

Global variables that are ignored:
\itemize{
\item \code{.Random.seed}
}
}
\examples{
\dontrun{
addTaskCallback(tracker_globals, name = ".GlobalEnv tracker")
}

}
